Fix(dashboard warning):  create a new array to destructure the _id property from the _doc property of each transaction.
newTransactionsArray transforms each MongoDB document in the original array into a new object where the _id property is a string.

The warning was: "only plain objects can be passed to Client Components from Server Components. Objects with toJSON methods are not supported. Convert it manually to a simple value before passing it to props."